+++ src/Resolvers/BrowserPreferredLocale.php
@@ -5,7 +5,7 @@ namespace Goodcat\L10n\Resolvers;
 use Goodcat\L10n\Routing\RouteTranslations;
 use Illuminate\Http\Request;
 
-class BrowserLocale implements PreferredLocaleResolver
+class BrowserPreferredLocale implements PreferredLocaleResolver
 {
     public function resolve(Request $request): ?string
     {